gcp/bigquery/environment_discovery/requirements.txt
google-cloud-bigquery
google-cloud-resource-manager
pandas